+++ script/blenderRCBPanel/__init__.py
@@ -34,7 +34,7 @@ from .blenderRCBPanel import (MyProperties,
                               WM_OT_Connect,
                               WM_OT_Configure,
                               WM_OT_ReachTarget,
-                              ModalOperator,
+                              #ModalOperator,
                               OBJECT_PT_robot_controller,
                               OT_OpenConfigurationFile,
                               ListItem,
@@ -51,7 +51,7 @@ classes = (
     WM_OT_Connect,
     WM_OT_Configure,
     WM_OT_ReachTarget,
-    ModalOperator,
+    #ModalOperator,
     OBJECT_PT_robot_controller,
     OT_OpenConfigurationFile,
     ListItem,
